[FB対応]index.tsxにてexportのとりまとめ
